+++ src/odm_validation/part_tables.py
@@ -1,18 +1,34 @@
 """Part-table definitions."""
 
+import logging
 from dataclasses import dataclass
+from enum import Enum
 from semver import Version
 from typing import Dict, List, Set
 
 import utils
-from utils import meta_get, meta_mark, meta_pop
-from versions import MapKind, get_mapping, has_mapping, is_compatible
+from versions import parse_version
 
 
 # type aliases
 Row = dict
 Dataset = List[Row]
 Schema = dict  # A Cerberus validation schema
+
+
+class MapKind(Enum):
+    TABLE = 0
+    ATTRIBUTE = 1
+    CATEGORY = 2
+
+
+# TODO: rename id to part_id?
+@dataclass(frozen=True)
+class Mapping:
+    kind: MapKind
+    id: str
+    table: str
+    meta: dict
 
 
 @dataclass(frozen=True)
@@ -68,6 +84,87 @@ TABLE = 'table'
 # other value constants
 MANDATORY = 'mandatory'
 NA = {'', 'NA', 'Not applicable'}
+
+# mapping
+V1_KIND_MAP = {
+    'tables': MapKind.TABLE,
+    'variables': MapKind.ATTRIBUTE,
+    'variableCategories': MapKind.CATEGORY,
+}
+
+
+def parse_row_version(row, field, default=None):
+    return parse_version(row.get(field), row.get('partID'), field, default)
+
+
+def is_compatible(part: dict, version: Version) -> bool:
+    # TODO: remove default for `firstReleased` when parts-v2 is complete
+    row = part
+    v1 = Version(major=1)
+    first = parse_row_version(row, 'firstReleased', default=v1)
+    last = parse_row_version(row, 'lastUpdated', default=first)
+    active: bool = row.get('status') == 'active'
+
+    # not (v < first) and ((v < last) or active)
+    v = version
+    if v.compare(first) < 0:
+        return False
+    if v.compare(last) < 0:
+        return True
+    return active
+
+
+def _get_original_key_val(part, key, val=None):
+    key_orig = part.get(key + _ORIGINAL_KEY)
+    val_orig = part.get(key + _ORIGINAL_VAL)
+    if key_orig:
+        key = key_orig
+    if val_orig:
+        val = val_orig
+    elif not val:
+        val = part.get(key)
+    return (key, val)
+
+
+def meta_mark(meta, part, key, val=None):
+    (key, val) = _get_original_key_val(part, key)
+    meta[key] = val
+    # print(f'meta marked ({key}, {val})')
+
+
+def meta_get(meta, part, key):
+    """returns `part[key]` and records the retrival in `meta`"""
+    val = part.get(key)
+    meta_mark(meta, part, key, val)
+    return val
+
+
+def meta_pop(meta, part, key):
+    """same as `meta_get` but also removes `key` from `part`."""
+    result = meta_get(meta, part, key)
+    del part[key]
+    return result
+
+
+def get_mapping(part: dict, version: Version) -> Mapping:
+    """Returns `None` when no mapping exists."""
+    if version.major != 1:
+        return
+    meta = {}
+    table = meta_get(meta, part, 'version1Table')
+    loc = meta_get(meta, part, 'version1Location')
+    kind = V1_KIND_MAP.get(loc)
+    if kind == MapKind.TABLE:
+        id = table
+    else:
+        id = meta_get(meta, part, 'version1Variable')
+    if not (kind and table and id):
+        return
+    return Mapping(kind=kind, id=id, table=table, meta=meta)
+
+
+def has_mapping(part: dict, version: Version) -> bool:
+    return get_mapping(part, version) is not None
 
 
 def table_required_field(table_name):
@@ -201,7 +298,7 @@ def transform_v2_to_v1(parts0: Dataset) -> (Dataset, dict):
 
         # TODO
         if mapping.kind == MapKind.CATEGORY:
-            error(f'{mapping.kind} is not yet implemented')
+            logging.error(f'{mapping.kind} is not yet implemented')
             p1 = p0
             continue
