+++ src/odm_validation/utils.py
@@ -1,4 +1,3 @@
-from logging import debug
 from typing import List
 
 import csv
@@ -25,42 +24,6 @@ def deep_update(src: dict, dst: dict):
             if len(src_list) > 0:
                 dst_list = dst[key]
                 dst_list += src_list
-
-
-def _get_original_key_val(part, key, val=None):
-    key_orig = part.get(key + '_original_key')
-    val_orig = part.get(key + '_original_val')
-    if key_orig:
-        key = key_orig
-    if val_orig:
-        val = val_orig
-    elif not val:
-        val = part.get(key)
-    return (key, val)
-
-
-def meta_mark(meta, part, key, val=None):
-    (key, val) = _get_original_key_val(part, key)
-    meta[key] = val
-    # print(f'meta marked ({key}, {val})')
-
-
-def meta_get(meta, part, key):
-    """returns `part[key]` and records the retrival in `meta`"""
-    val = part.get(key)
-    meta_mark(meta, part, key, val)
-    return val
-
-
-def meta_pop(meta, part, key):
-    """same as `meta_get` but also removes `key` from `part`."""
-    result = meta_get(meta, part, key)
-    del part[key]
-    return result
-
-
-# def meta_set(meta, part, key):
-#     """returns `part[key]` and records the retrival in `meta`"""
 
 
 def import_dataset(fileName: str) -> List[dict]:
