+++ src/odm_validation/versions.py
@@ -1,39 +1,11 @@
 import os
 import re
 import toml
-from dataclasses import dataclass
-from enum import Enum
 from importlib import metadata
 from logging import warning
 from os import path
 from semver import Version
 from typing import Optional, Tuple
-
-from utils import meta_get
-
-
-# __all__ = [
-#     '__version__',
-#     'Version',
-#     'get_mapping',
-#     'is_compatible',
-#     'parse_version',
-# ]
-
-
-class MapKind(Enum):
-    TABLE = 0
-    ATTRIBUTE = 1
-    CATEGORY = 2
-
-
-# TODO: rename id to part_id?
-@dataclass(frozen=True)
-class Mapping:
-    kind: MapKind
-    id: str
-    table: str
-    meta: dict
 
 
 def _get_package_version():
@@ -59,13 +31,6 @@ BASEVERSION = re.compile(
     """,
     re.VERBOSE,
 )
-
-V1_KIND_MAP = {
-    'tables': MapKind.TABLE,
-    'variables': MapKind.ATTRIBUTE,
-    'variableCategories': MapKind.CATEGORY,
-}
-
 
 def _coerce(version: str) -> Tuple[Version, Optional[str]]:
     """
@@ -116,45 +81,3 @@ def parse_version(version: str, id='', label='', default: Version = None
         (result, _) = _coerce(version)
         warning(correctionMsg.format(result))
         return result
-
-
-def parse_row_version(row, field, default=None):
-    return parse_version(row.get(field), row.get('partID'), field, default)
-
-
-def is_compatible(part: dict, version: Version) -> bool:
-    # TODO: remove default for `firstReleased` when parts-v2 is complete
-    row = part
-    v1 = Version(major=1)
-    first = parse_row_version(row, 'firstReleased', default=v1)
-    last = parse_row_version(row, 'lastUpdated', default=first)
-    active: bool = row.get('status') == 'active'
-
-    # not (v < first) and ((v < last) or active)
-    v = version
-    if v.compare(first) < 0:
-        return False
-    if v.compare(last) < 0:
-        return True
-    return active
-
-
-def get_mapping(part: dict, version: Version) -> Mapping:
-    """Returns `None` when no mapping exists."""
-    if version.major != 1:
-        return
-    meta = {}
-    table = meta_get(meta, part, 'version1Table')
-    loc = meta_get(meta, part, 'version1Location')
-    kind = V1_KIND_MAP.get(loc)
-    if kind == MapKind.TABLE:
-        id = table
-    else:
-        id = meta_get(meta, part, 'version1Variable')
-    if not (kind and table and id):
-        return
-    return Mapping(kind=kind, id=id, table=table, meta=meta)
-
-
-def has_mapping(part: dict, version: Version) -> bool:
-    return get_mapping(part, version) is not None
